requirements.txt
streamlit>=1.32.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-docx>=1.1.0
python-pptx>=0.6.23
anthropic>=0.25.0
